refactor: move the celcius and farenheit functions to its own directory
