+++ src/constants.js
@@ -1,5 +1,8 @@
 /** @type {number} conversion factor from farenheit to celcius */
 export const CelciusFactor = 5 / 9;
 
+/** @type {number} conversion factor from celcius to farenheit */
+export const FarenheitFactor = 9 / 5;
+
 /** @type {number} Offset between celcius and farenheit zero value */
 export const TemperatureOffset = 32;
